Moves toward bbrl_gym keeping my_gym
