+++ .gitignore
@@ -1,0 +1,3 @@
+*.egg-info
+__pycache__
+dist/
