--- my_gym/envs/__init__.py
@@ -1,7 +1,0 @@
-from my_gym.envs.continuous_cartpole import ContinuousCartPoleEnv
-from my_gym.envs.line_mdp import LineMDPEnv
-from my_gym.envs.continuous_line_mdp import ContinuousLineMDPEnv
-from my_gym.envs.continuous_2D_mdp import Continuous2DMDPEnv
-from my_gym.envs.rocket_lander import RocketLanderEnv
-from my_gym.envs.maze_mdp import MazeMDPEnv
-from my_gym.envs.debug_env import DebugVEnv
