+++ setup.cfg
@@ -3,14 +3,14 @@
 license_file = LICENSE
 
 [pytype]
-inputs = my_gym
+inputs = bbrl_gym
 
 [flake8]
 ignore = W503,W504,E203,E231  # line breaks before and after binary operators
 # Ignore import not used when aliases are defined
 per-file-ignores =
-	./my_gym/__init__.py:F401
-	./my_gym/env/__init__.py:E402
+	./bbrl_gym/__init__.py:F401
+	./bbrl_gym/env/__init__.py:E402
 exclude =
     # No need to traverse our git directory
     .git,
@@ -30,4 +30,4 @@ max-line-length = 127
 [isort]
 profile = black
 line_length = 127
-src_paths = my_gym
+src_paths = bbrl_gym
