+++ tests/test_maze.py
@@ -1,10 +1,10 @@
 import gym
-import my_gym
+import bbrl_gym
 import numpy as np
 from typing import Tuple, List
 
 
-from my_gym.envs.maze_mdp import MazeMDPEnv
+from bbrl_gym.envs.maze_mdp import MazeMDPEnv
 
 
 def get_policy_from_v(mdp: MazeMDPEnv, v: np.ndarray) -> np.ndarray:
@@ -13,16 +13,14 @@ def get_policy_from_v(mdp: MazeMDPEnv, v: np.ndarray) -> np.ndarray:
     for x in range(mdp.nb_states):  # for each state x
         # Compute the value of the state x for each action u of the MDP action space
         v_temp = []
-        for u in mdp.action_space.actions:
-            if x not in mdp.terminal_states:
+        if x not in mdp.terminal_states:
+            for u in mdp.action_space.actions:
                 # Process sum of the values of the neighbouring states
                 summ = 0
                 for y in range(mdp.nb_states):
                     summ = summ + mdp.P[x, u, y] * v[y]
                 v_temp.append(mdp.r[x, u] + mdp.gamma * summ)
-            else:  # if the state is final, then we only take the reward into account
-                v_temp.append(mdp.r[x, u])
-        policy[x] = np.argmax(v_temp)
+            policy[x] = np.argmax(v_temp)
     return policy
 
 
@@ -36,28 +34,26 @@ def value_iteration_v(
     stop = False
 
     if render:
-        mdp.init_draw("Value iteration V")
+        mdp.init_draw("Value iteration V", mode="rgb_array")
 
     while not stop:
         v_old = v.copy()
         if render:
-            mdp.draw_v(v, title="Value iteration V")
+            mdp.draw_v(v, title="Value iteration V", mode="rgb_array")
 
         for x in range(mdp.nb_states):  # for each state x
             # Compute the value of the state x for each action u of the MDP action space
-            v_temp = []
-            for u in mdp.action_space.actions:
-                if x not in mdp.terminal_states:
+            if x not in mdp.terminal_states:
+                v_temp = []
+                for u in mdp.action_space.actions:
                     # Process sum of the values of the neighbouring states
                     summ = 0
                     for y in range(mdp.nb_states):
                         summ = summ + mdp.P[x, u, y] * v_old[y]
                     v_temp.append(mdp.r[x, u] + mdp.gamma * summ)
-                else:  # if the state is final, then we only take the reward into account
-                    v_temp.append(mdp.r[x, u])
 
-                    # Select the highest state value among those computed
-            v[x] = np.max(v_temp)
+                # Select the highest state value among those computed
+                v[x] = np.max(v_temp)
 
         # Test if convergence has been reached
         if (np.linalg.norm(v - v_old)) < 0.01:
@@ -66,7 +62,7 @@ def value_iteration_v(
 
     if render:
         policy = get_policy_from_v(mdp, v)
-        mdp.draw_v_pi(v, policy, title="Value iteration V")
+        mdp.draw_v_pi(v, policy, title="Value iteration V", mode="rgb_array")
 
     return v, v_list
